feat: customer details, update form, hero img

+++ src/components/CustomerList.jsx
@@ -1,33 +1,38 @@
 import { useDispatch, useSelector } from "react-redux";
 import { deleteCustomer } from "../features/customersSlice";
+import { Outlet, useNavigate } from "react-router-dom";
 
 const CustomerList = () => {
   const dispatch = useDispatch();
+  const navigate = useNavigate();
   const customers = useSelector((state) => state.customers);
 
   const handleDelete = (id) => {
-    confirm("Are you sure, Do you want to delete record?");
-    dispatch(deleteCustomer(id));
+    const isConfirm = confirm("Are you sure, Do you want to delete record?");
+    if (isConfirm) dispatch(deleteCustomer(id));
+  };
+
+  const handleUpdate = (id) => {
+    navigate(`customer/${id}`);
   };
   return (
-    <>
-      <div className="w-1/2 p-4">
-        <h2 className="text-3xl font-semibold py-2">CustomerList</h2>
-
+    <div className="md:flex">
+      <div className="w-1/2 p-4 text-center flex flex-col border-r-2 space-y-3">
+        <h2 className="text-3xl  font-semibold py-2">Customer List</h2>
         {customers.length > 0 ? (
-          <table className="divide-y divide-slate-900">
+          <table className="divide-y divide-slate-900 bg-slate-100 shadow-md">
             <thead>
               <tr className="bg-slate-200">
-                <th className="px-7 py-3 font-medium uppercase text-lg">
-                  Name
+                <th className="px-7 py-3 font-medium capitalize text-lg">
+                  Customer Name
                 </th>
-                <th className="px-7 py-3 font-medium uppercase text-lg">
-                  No of Purchase
+                <th className="px-7 py-3 font-medium capitalize text-lg">
+                  No. of Purchase Items
                 </th>
-                <th className="px-7 py-3 font-medium uppercase text-lg">
-                  amount
+                <th className="px-7 py-3 font-medium capitalize text-lg">
+                  Amount
                 </th>
-                <th className="px-4 py-3 font-medium uppercase text-lg">
+                <th className="px-4 py-3 font-medium capitalize text-lg">
                   Action
                 </th>
               </tr>
@@ -36,17 +41,23 @@ const CustomerList = () => {
               {customers.map((customer) => {
                 return (
                   <tr key={customer.id}>
-                    <td className="px-7 py-4 text-lg">
+                    <td
+                      className="px-7 py-2 text-lg cursor-pointer hover:bg-slate-200"
+                      onClick={() => navigate(`customer/edit/${customer.id}`)}
+                    >
                       <div>{customer.name}</div>
                     </td>
-                    <td className="px-7 py-4 text-lg">
+                    <td className="px-7 py-2 text-lg">
                       <div>{customer.noOfPurchase}</div>
                     </td>
-                    <td className="px-7 py-4 text-lg">
+                    <td className="px-7 py-2 text-lg">
                       <div>{customer.amount}</div>
                     </td>
-                    <td className="px-7 py-4 text-lg">
-                      <button className="bg-slate-500 hover:bg-slate-700 text-white mr-2 px-2 py-1 sm:px-2 sm:py-1 md:px-3 md:py-2 rounded-md">
+                    <td className="px-7 py-2 text-lg">
+                      <button
+                        onClick={() => handleUpdate(customer.id)}
+                        className="bg-slate-500 hover:bg-slate-700 text-white mr-2 px-2 py-1 sm:px-2 sm:py-1 md:px-3 md:py-2 rounded-md"
+                      >
                         Edit
                       </button>
                       <button
@@ -67,7 +78,10 @@ const CustomerList = () => {
           </div>
         )}
       </div>
-    </>
+      <div className="w-1/2 p-4">
+        <Outlet />
+      </div>
+    </div>
   );
 };
